+++ src/controllers/cart.controller.js
@@ -2,7 +2,8 @@ const cartService = require("../services/cart.service");
 
 const insertCart = async (req, res) => {
   try {
-    const { userId, productId, productCount } = req.body;
+    const userId = req.user.id;
+    const { productId, productCount } = req.body;
 
     await cartService.insertCart(userId, productId, productCount);
     res.status(201).end();
