+++ src/routes/cart.router.js
@@ -2,8 +2,9 @@ const express = require("express");
 const cartRouter = express.Router();
 
 const cartController = require("../controllers/cart.controller");
+const { loginRequired } = require("../utils/auth");
 
-cartRouter.post("/insertCart", cartController.insertCart);
-cartRouter.delete("/deleteCart", cartController.deleteCart);
+cartRouter.post("/insertCart", loginRequired, cartController.insertCart);
+cartRouter.delete("/deleteCart", loginRequired, cartController.deleteCart);
 
 module.exports = { cartRouter };
