+++ src/routes/index.js
@@ -2,7 +2,7 @@ const express = require("express");
 
 const { cartRouter } = require("./cart.router");
 
-const { userRouter } = require("./userRouter");
+const { userRouter } = require("./user.router");
 
 const router = express.Router();
 router.use("/cart", cartRouter);
