+++ src/routes/user.router.js
@@ -1,0 +1,8 @@
+const express = require("express");
+const userController = require("../controllers/user.controllers");
+const userRouter = express.Router();
+
+userRouter.post("/signup", userController.signUp);
+userRouter.post("/signin", userController.signIn);
+
+module.exports = { userRouter };
